@@ -1,27 +1,24 @@
 {
-  "name": "Sellify",
+  "name": "sellify",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@apollo/react-hooks": "^4.0.0",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
-    "autoprefixer": "^10.4.7",
+    "apollo-boost": "^0.4.9",
     "axios": "^0.27.2",
-    "cssnano": "^5.1.12",
-    "gulp": "^4.0.2",
-    "gulp-cli": "^2.3.0",
-    "gulp-postcss": "^9.0.1",
+    "bootstrap": "^5.1.3",
+    "jquery": "^3.6.0",
     "js-cookie": "^3.0.1",
-    "postcss": "^8.4.14",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.4.0",
-    "react-dom": "^17.0.2",
-    "react-hook": "^0.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^8.0.2",
-    "react-router-bootstrap": "^0.26.1",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
