@@ -1,22 +1,25 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+// import { Provider } from 'react-redux';
+//import ReactDOM from 'react-dom/client';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
-import { HashRouter } from 'react-router-dom';
 
-// import './dist/css/main.css';
-// import { HashRouter } from 'react-router-dom';
+// import store from './store';
 
-import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-  <HashRouter>
+
+// ReactDOM.render(
+//   <React.StrictMode>
+//     <App />
+//   </React.StrictMode>,
+//   document.getElementById('root')
+// );
+
+const root = createRoot(document.getElementById('root'))
+
+root.render(
+  <React.StrictMode>
     <App />
-  </HashRouter>,
-  document.getElementById('root')
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+  </React.StrictMode>
+)
